refactor: extract Include-path parsing from LC006 into shared IncludePathParser — LC045 (missing include) needs the same lambda/string/ThenInclude path parsing
